Adding Driver Rating to DB

+++ src/main/java/com/mycompany/gocheeta/rest/Users.java
@@ -17,7 +17,7 @@ public class Users {
     private int noOfTrips;
     private String status;
     private int noOfVehicles;
-    private int rating;
+    private double rating;
     
     public Users() {
         this.email = "";
@@ -31,7 +31,7 @@ public class Users {
         this.rating = 0;
     }
 
-    public Users(String email, String password, String name, String branch, String telephone, int noOfTrips, String status, int noOfVehicles, int rating) {
+    public Users(String email, String password, String name, String branch, String telephone, int noOfTrips, String status, int noOfVehicles, double rating) {
         this.email = email;
         this.password = password;
         this.name = name;
@@ -107,11 +107,11 @@ public class Users {
         this.noOfVehicles = noOfVehicles;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
